Fix up comments from review
- Truncate values to double before passing to JS
- Only call RuntimeMemory.getInstance() once
- Chain builder calls

+++ web/client-api/src/main/java/io/deephaven/web/client/api/JsWorkerHeapInfo.java
@@ -15,17 +15,17 @@ public class JsWorkerHeapInfo {
     }
 
     @JsProperty
-    public long getMaximumHeapSize() {
+    public double getMaximumHeapSize() {
         return maximumHeapSize;
     }
 
     @JsProperty
-    public long getFreeMemory() {
+    public double getFreeMemory() {
         return freeMemory;
     }
 
     @JsProperty
-    public long getTotalHeapSize() {
+    public double getTotalHeapSize() {
         return totalHeapSize;
     }
 }
